Stucture html, create script and styles
I simply forgot to commit bit by bit, and I did a big part of the
exercice which is to display the library, so going through the library
array and creating cards with the the content inside. I had to review
some of the functions to manipulate the DOM (it's hard this part).