Create first form elements inside dialog and style
I want to practice using dialogs with putting the form for creating new
books inside one. I've just layed the basic elements to display the
dialog when the button is clicked, and of course some styling on the
dialog. I still need to figure out how to use the inputs results to
create a new book.