Add the toggleBeenRead prototype func and button
I finish the exercice using the same method for the remove button, with
simply displaying again the full library (i understand it's not the best
way because you reload everything and it would be better to just target
the node you want to remove/toggle). For that I add the toggleBennRead
function on the Book.prototype.